+ Adds new field named 'internalFormat in Texture node. ~ Updates LightModel node by adding INT24 support of 'shadowMapType' field and by modifying its default value to INT32. + Adds fields named 'distanceAttenuation' and 'smooth' to PointStyle node. + Adds a new node named 'FrameBuffer' representing the current framebuffer. ~ Updates Program node to override the current auto-generated GLSL program.

+++ core/vgd/src/node/LightModel.cpp
@@ -70,7 +70,7 @@ LightModel::LightModel( const std::string nodeName ) :
 void LightModel::setToDefaults( void )
 {
 	SingleAttribute::setToDefaults();
-	setShadowMapType( INT16 );
+	setShadowMapType( INT32 );
 	setIlluminationInShadow( 0.4f );
 	setSamplingSize( 1.0 );
 }
